test: sample with token generate api
