+++ turbo.json
@@ -4,7 +4,7 @@
   "pipeline": {
     "build": {
       "dependsOn": ["^build"],
-      "env": ["NEXT_PUBLIC_GATEWAYS"],
+      "env": ["NEXT_PUBLIC_GATEWAYS", "NEXT_PUBLIC_APP_TOKEN"],
       "outputs": [".next/**", "!.next/cache/**"]
     },
     "lint": {
